Facebook Qt for Symbian OS
Facebook Qt formally Facebook Touch Qt is a web Facebook client for Symbian Belle and higher devices.

This is a old project so please excuse the coding style  in places.

The App version for this code is v1.95 EOL.

**When importing the project you will receive a message regarding qmlapplicationviewer class, ignore this message**

Feel free to change the App and publish your own builds but please leave the original credits in the About Page.